test: migrate asar specs to main runner
